Handle premultiplied images with 'marginal' pixels
Some premultiplied images have a few pixel values
that are just outside of the correct range, which
results in the image being processed as if it was
not premultiplied. Fixed alpha checking to allow
for these images and made sure conversion does not
allow pixels to get out of range.